Disable submit button if merge will fail
This feature aims to test the mergeability of each change tagging it
according to the result (true or false).
When the change page is load, the test is triggered and the sha1 of
the dest branch (tested against) is stored within the change table.
The submit button will be enabled or disabled according to the result.
The test is triggered only if it’s outdated.
As the branch evolves the sha1 stored at the moment of the test
becomes different from the last one merged into the branch, meaning
the test is outdated.

Bug: issue 381
Change-Id: I8bdb26db6fc35067b8c14b987984cfd80745f3d3