TopicReviews: Added TopicControl for Topic access management
This commit aims to implement the access control management for a user
accessing a single topic. It also introduces some changes in the
ChangeControl behaviour regarding the changes belonging to a Topic.

TODO: The "canSubmit" for Topics needs to be implemented using the new
Prolog implementation. Here the implementation is the previous one.

Change-Id: If6d9f11c62a6833b92f477ba52c15f6c54ca8eee