TopicReviews: Added main Topic Screen
Added Topic Screen to handle the different actions over topics. It will
follow the Change Screen distribution, but with some changes. The patch
set list will be replaced by a list of change sets, each one of them may
display a list with several changes belonging to the topic.
It will be possible to navigate through the change list using the usual
shortcut keys.

Change-Id: I9c5e751f5aad4d8aae9adb9d292e0e7071f350fa