Enhance merge conflict handling for topics
If a merge conflict occurred for a change associated with a topic
being merged the user was give very little feedback that a conflict
had occured. A message was appended to the change and the topic was
left in the merge pending state. This fix adds a comment to the topic
indicating that a merge conflict has occurred for one of the changes
associated with the topic, the message is also provided in a "Submit Failed"
dialog displayed to the user. The topic status is reset to "New". This
behaviour is consistent with the error produced when change not associated
with a topic has merge conflicts.

Change-Id: I6d816a43458d198f1776d4f6cfcb9e26ab6cd350